Update device_info.py
We need to update the NETCONF ports and URLs for the Always-On IOS XE on CSR Recommended Code device in the device_info.py file. The DevNet Always-On Sandbox APIC-EM sandbox is no longer in the catalog and neither is the IOS XE Programmability Sandbox Device. The password for the Always-On Sandbox ACI APIC sandbox, which is now the ACI Simulator AlwaysOn - V5 sandbox, needs to be changed as well.

Once the changes are made, we can run the toolkit_getting_started.py and toolkit_new_policy.py files and see the results in the IDE, although I'm not seeing the Controller Policy change in the GUI (I might be looking in the wrong place, though)